langchain/requirements.txt
# requirements.txt

#langchain==0.3.4
langchain==0.1.11 
langchain-core==0.1.29
langchain-openai==0.2.3
langchain-community==0.3.3
faiss-cpu==1.9.0
openai==1.55.3 
httpx==0.27.2   
tenacity==9.1.2
python-dotenv==1.0.1  
langchain-google-genai==0.0.6
langchain-chroma==0.1.4

# PDF + OCR + Image processing
pypdf==4.3.1
pdfplumber==0.11.5
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0

#chromadb==0.5.13
tqdm==4.66.4

pytesseract
#chromadb-0.5.23
chromadb-0.5.17

